Added save function. Further refactorization of CLI.

+++ business-planner/src/session/mod.rs
@@ -1,12 +1,12 @@
 use serde::{Deserialize, Serialize};
-use std::{fs, path::{Path, PathBuf}};
+use std::{fs, path::{self, Path, PathBuf}};
 
 use crate::errors::session::{LoadSessionError, SaveSessionError};
 
 #[derive(Default)]
 pub struct Session {
-    last_save_location: Option<PathBuf>,
-    data: SessionData,
+    pub last_save_location: Option<PathBuf>,
+    pub data: SessionData,
 }
 
 #[derive(Debug, Serialize, Deserialize, PartialEq)]
@@ -22,18 +22,22 @@ impl Default for SessionData {
     }
 }
 
-
 impl Session {
-    pub fn save_to_last_save_location(&self) -> Result<(), SaveSessionError> {
+    pub fn save_to_last_save_location(&self, overwrite: bool) -> Result<(), SaveSessionError> {
         match &self.last_save_location {
-            Some(path) => self.save_to_location(path),
+            Some(path) => self.save_to_location(path, overwrite),
             None => Err(SaveSessionError::UndefinedSavePath),
         }
     }
 
-    pub fn save_to_location(&self, path: &Path) -> Result<(), SaveSessionError> {
-        let serialized = serde_xml_rs::to_string(&self.data)?;
+    pub fn save_to_location(&self, path: &Path, overwrite: bool) -> Result<(), SaveSessionError> {
+        let serialized: String = serde_xml_rs::to_string(&self.data)?;
+        if path.exists() && !overwrite {
+            return Err(SaveSessionError::FileExists)
+        }
+
         fs::write(path, serialized)?;
+        
         Ok(())
     }
 }
@@ -45,8 +49,9 @@ pub fn create() -> Session {
 pub fn load(path: PathBuf) -> Result<Session, LoadSessionError> {
     let serialized_session_data = fs::read_to_string(&path)?;
     let session_data = serde_xml_rs::from_str(&serialized_session_data)?;
+    let path = path::absolute(path).ok();
     Ok(Session {
-        last_save_location: Some(path),
+        last_save_location: path,
         data: session_data,
     })
 }
